+++ www/config/constants.php
@@ -1,0 +1,14 @@
+<?php
+
+    // project name
+    //
+    $m_name = "monoto";
+
+    // Github Links
+    //
+    $m_githubProjectLink = "https://github.com/yafp/monoto";
+    $m_githubIssueLink = "https://github.com/yafp/monoto/issues";
+    $m_githubChangelogLink = "https://github.com/yafp/monoto/blob/master/docs/CHANGELOG.md";
+    $m_githubWikiLink = "https://github.com/yafp/monoto/wiki";
+
+?>