+++ www/inc/genericHeaderCSS.php
@@ -4,17 +4,21 @@
 <!-- 4.2.1 -->
 <link rel="stylesheet" type="text/css" href="js/bootstrap/4.2.1/css/bootstrap.min.css">
 
-<!-- Bootstrap-select -->
-<!-- 1.13.7 -->
-<link rel="stylesheet" type="text/css" href="js/bootstrap-select/1.13.7/dist/css/bootstrap-select.min.css">
 
 <!-- DataTables -->
 <!-- 1.10.18 -->
 <link rel="stylesheet" type="text/css" href="js/datatables/1.10.18/css/jquery.dataTables.min.css" />
 
+
 <!-- Font Awesome -->
 <!-- 5.7.0 -->
 <link rel="stylesheet" type="text/css" href="js/FontAwesome/5.7.0/css/fontawesome.min.css">
+
+
+<!-- select2-->
+<!-- 4.0.5 -->
+<link rel="stylesheet" type="text/css" href="js/select2/4.0.5/dist/css/select2.min.css">
+
 
 <!-- Monoto specific css -->
 <link rel="stylesheet" type="text/css" href="css/monoto/general.css" title="default" />