+++ docs/CHANGELOG.md
@@ -75,6 +75,7 @@
   * [#284] Added .htaccess for some important folders
   * [#291] Localize datatable
   * [#290] Bootstrap <select>
+  * [#296] Switch from bootstrap-select to select2
 
 ###  ```Removed```
 * Notes